Update react peerDependencies
This project works just fine with React 17. However, new users using the latest node and npm will encounter an error that prohibits installation of this project due to peerDependencies.  Users must do `npm install --force` or `npm install --use-legacy-deps` in order to force the installation.  This is not ideal.  This update resolves the issue.